requirements.txt
blinker==1.8.2
click==8.1.8
dnspython==2.6.1
email-validator==2.3.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
pkg_resources==0.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.43
typing_extensions==4.13.2
Werkzeug==3.0.6
WTForms==3.1.2
zipp==3.20.2
